+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added `PixState::smooth` and `PixState::no_smooth` to toggle anti-alias
+  drawing of shapes.
 
 ### Changed
 
@@ -42,8 +44,8 @@ All notable changes to this project will be documented in this file.
 
 ### Breaking
 
-- Disabled audio playback by default on startup. To queue and play audio you must first call
-  `PixState::resume_audio`.
+- Disabled audio playback by default on startup. To queue and play audio you
+  must first call `PixState::resume_audio`.
 
 ## [0.5.2] - 2021-12-13